// models/accounting/vendors/vendorsDetails.ts
export interface VendorsDetails {
    id: number;                        
    [x: string]: any;

  name: string
  contact_name: string
  email: string
  phone_number: string
  address_line1: string
  address_line2: string
  state: string
  city: string
  zip_code: string
  country: string
  business_hours: string
  line_of_business: string
  status: string
  account_number: string
  bank_name: string
  routing_number: string
  }